requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers==4.36.2
accelerate==0.25.0
scikit-learn==1.3.2
matplotlib==3.8.2
streamlit==1.29.0
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.1.0
numpy==1.26.2
tqdm==4.66.1
sentencepiece==0.1.99
bitsandbytes==0.41.3
protobuf==4.25.1 